visualassistai/App-for-Visually-Impaired-Individuals-main/requirements.txt
google.generativeai
streamlit
Pillow
gtts
matplotlib
